+++ app/src/main/java/com/makebodywell/bodywell/model/DailyData.kt
@@ -5,7 +5,7 @@ data class DailyData(
     var foodGoal: Int = 0,
     var waterGoal: Int = 0,
     var exerciseGoal: Int = 0,
-    var bodyGoal: Int = 0,
+    var bodyGoal: Double = 0.0,
     var sleepGoal: Int = 0,
     var drugGoal: Int = 0,
     var regDate: String = ""
